LPD-51312 Split frontend-js-state-web in UI and non-UI parts
Note that we need to put State in its own submodule (impl) so that at
runtime it is always obtained from a single bundle wherever it is used.
Otherwise we end up having more than one State and React enters infinite
loops.
