+++ modules/apps/frontend-js/frontend-js-state-web/src/main/resources/META-INF/resources/impl/deepFreeze.ts
